django_elasticsearch/requirements.txt

asgiref==3.8.1
django==5.1.7
sqlparse==0.5.3
djangorestframework==3.15.2

certifi==2025.1.31 
django-elasticsearch-dsl==8.0 
elastic-transport==8.17.0 
elasticsearch==8.17.2 
elasticsearch-dsl==8.17.1 
python-dateutil==2.9.0.post0 
six==1.17.0 
typing-extensions==4.12.2 
urllib3==2.3.0
